Moved the simulation type and discretization type properties from MeshDescritizer class to PhysicsContinuum class.
