+++ Channel_Flow_Solver/PhysicsContinuum.cpp
@@ -46,4 +46,35 @@ void PhysicsContinuum::setDiffusionCoefficient(double value) {
     this->diffusionCoefficient = value;
 }
 
+PhysicsContinuum::SimulationType PhysicsContinuum::getsimulationType() const {
+    return simulationType;
+}
+
+void PhysicsContinuum::setsimulationType(SimulationType type) {
+    this->simulationType = type;
+    if(type == SimulationType::DIFFUSION){
+        diffusionEnabled = true;
+    }else if (type == SimulationType::CONVECTION_DIFFUSION){
+        convectionEnabled = true;
+        diffusionEnabled = true;
+    }
+}
+
+PhysicsContinuum::ConvectionDifferencingScheme PhysicsContinuum::getConvectionDifferencingScheme() const {
+    return differencingScheme;
+}
+
+void PhysicsContinuum::setConvectionDifferencingScheme(ConvectionDifferencingScheme differencingScheme) {
+    this->differencingScheme = differencingScheme;
+}
+
+Material* PhysicsContinuum::getMaterial() const {
+    return material;
+}
+
+void PhysicsContinuum::setMaterial(Material* mat) {
+    this->material = material;
+}
+
+
