+++ Channel_Flow_Solver/PhysicsContinuum.h
@@ -14,8 +14,15 @@
 #ifndef PHYSICSCONTINUUM_H
 #define PHYSICSCONTINUUM_H
 
+#include "Material.h"
+
 class PhysicsContinuum {
+    
 public:
+    
+    enum SimulationType {DIFFUSION, CONVECTION_DIFFUSION}; 
+    enum ConvectionDifferencingScheme {CENTRAL, FIRST_ORDER_UPWIND};  
+    
     PhysicsContinuum();
     PhysicsContinuum(const PhysicsContinuum& orig);
     virtual ~PhysicsContinuum();
@@ -27,10 +34,22 @@ public:
     double getDensity() const;
     void setDensity(double value);
     
+    SimulationType getsimulationType() const;
+    void setsimulationType(SimulationType type);
+    ConvectionDifferencingScheme getConvectionDifferencingScheme() const;
+    void setConvectionDifferencingScheme(ConvectionDifferencingScheme differencingScheme);  
+    Material* getMaterial() const;
+    void setMaterial(Material* mat);
+    
 private:
     double thermalConductivity;
     double diffusionCoefficient;
     double density;
+    
+    SimulationType simulationType;
+    ConvectionDifferencingScheme differencingScheme;    
+    bool diffusionEnabled, convectionEnabled;
+    Material* material;
 };
 
 #endif /* PHYSICSCONTINUUM_H */
